Implement training kNN on RBM residuals
To run, 'make res' and then 'bin/res'

DO NOT MERGE THIS BRANCH INTO MASTER. kNN and RBM main()s have been commented out
and even if commented back in, things will probably break.

+++ src/utils.hpp
@@ -31,7 +31,7 @@ struct dataPoint {
     int userID;
     int movieID;
     int date;
-    int value;
+    float value;
 
     dataPoint() {
         userID = 0;
@@ -40,7 +40,7 @@ struct dataPoint {
         value = 0;
     }
 
-    dataPoint(int a, int b, int c, int d) :
+    dataPoint(int a, int b, int c, float d) :
         userID(a), movieID(b), date(c), value(d) {
     }
